+++ packages/react-wallet-kit/src/components/auth/wallet/utils.ts
@@ -1,0 +1,13 @@
+import type { WalletProvider } from "@turnkey/core";
+
+/** @internal */
+export const canDisconnect = (
+  provider: WalletProvider,
+  shouldShowDisconnect?: boolean,
+) => {
+  return (
+    shouldShowDisconnect &&
+    provider.connectedAddresses &&
+    provider.connectedAddresses.length > 0
+  );
+};
