+++ packages/react-wallet-kit/src/components/design/Svg.tsx
@@ -210,3 +210,3056 @@ export function SolanaLogo(props: SVGProps) {
     </svg>
   );
 }
+
+export function QRCodeIcon(props: SVGProps) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      xmlnsXlink="http://www.w3.org/1999/xlink"
+      width={400}
+      height={400}
+      viewBox="0 0 1000 1000"
+      enableBackground="new 0 0 400 400"
+      xmlSpace="preserve"
+      {...props}
+    >
+      <rect
+        x={0}
+        y={0}
+        width={1000}
+        height={1000}
+        fill="rgb(255,255,255)"
+        opacity={1}
+      />
+      <polygon
+        points="0,225 0,0 225,0 225,50, 50,50 50,225"
+        fill="rgb(0,0,0)"
+        opacity={1}
+      />
+      <circle cx={25} cy={225} r={25} fill="rgb(0,0,0)" opacity={1} />
+      <circle cx={225} cy={25} r={25} fill="rgb(0,0,0)" opacity={1} />
+      <polygon
+        points="775,0 1000,0 1000,225 950,225, 950,50 775,50"
+        fill="rgb(0,0,0)"
+        opacity={1}
+      />
+      <circle cx={775} cy={25} r={25} fill="rgb(0,0,0)" opacity={1} />
+      <circle cx={975} cy={225} r={25} fill="rgb(0,0,0)" opacity={1} />
+      <polygon
+        points="1000,775 1000,1000 775,1000 775,950, 950,950 950,775"
+        fill="rgb(0,0,0)"
+        opacity={1}
+      />
+      <circle cx={975} cy={775} r={25} fill="rgb(0,0,0)" opacity={1} />
+      <circle cx={775} cy={975} r={25} fill="rgb(0,0,0)" opacity={1} />
+      <polygon
+        points="0,775 0,1000 225,1000 225,950, 50,950 50,775"
+        fill="rgb(0,0,0)"
+        opacity={1}
+      />
+      <circle cx={25} cy={775} r={25} fill="rgb(0,0,0)" opacity={1} />
+      <circle cx={225} cy={975} r={25} fill="rgb(0,0,0)" opacity={1} />
+      <g transform="translate(100,100) scale(2.0,2.0)">
+        <svg
+          width={400}
+          height={400}
+          viewBox="0 0 580 580"
+          enableBackground="new 0 0 400 400"
+          xmlSpace="preserve"
+          {...props}
+        >
+          <rect
+            x={0}
+            y={0}
+            width={580}
+            height={580}
+            fill="rgb(255,255,255)"
+            opacity={1}
+          />
+          <rect
+            x={180}
+            y={0}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={0}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={0}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={20}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={20}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={20}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={340}
+            y={20}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={40}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={40}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={40}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={40}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={280}
+            y={40}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={320}
+            y={40}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={340}
+            y={40}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={360}
+            y={40}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={380}
+            y={40}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={40}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={60}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={60}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={220}
+            y={60}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={60}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={280}
+            y={60}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={60}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={80}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={80}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={80}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={280}
+            y={80}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={80}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={80}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={100}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={100}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={100}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={100}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={100}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={340}
+            y={100}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={360}
+            y={100}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={120}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={120}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={120}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={280}
+            y={120}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={320}
+            y={120}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={360}
+            y={120}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={120}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={140}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={140}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={140}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={140}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={360}
+            y={140}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={380}
+            y={140}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={140}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={0}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={40}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={60}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={80}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={100}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={120}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={280}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={320}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={360}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={440}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={460}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={500}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={520}
+            y={160}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={0}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={20}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={40}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={100}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={140}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={220}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={360}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={420}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={460}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={500}
+            y={180}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={0}
+            y={200}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={60}
+            y={200}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={80}
+            y={200}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={100}
+            y={200}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={120}
+            y={200}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={200}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={220}
+            y={200}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={200}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={280}
+            y={200}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={200}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={200}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={520}
+            y={200}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={540}
+            y={200}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={20}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={40}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={100}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={320}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={380}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={420}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={440}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={460}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={500}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={540}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={560}
+            y={220}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={0}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={60}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={80}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={100}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={120}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={140}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={280}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={440}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={240}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={0}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={40}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={60}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={140}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={220}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={280}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={320}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={360}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={440}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={500}
+            y={260}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={40}
+            y={280}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={60}
+            y={280}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={80}
+            y={280}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={120}
+            y={280}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={140}
+            y={280}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={280}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={280}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={280}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={280}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={340}
+            y={280}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={520}
+            y={280}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={560}
+            y={280}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={0}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={20}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={40}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={100}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={280}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={320}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={380}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={440}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={500}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={540}
+            y={300}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={20}
+            y={320}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={100}
+            y={320}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={120}
+            y={320}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={140}
+            y={320}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={320}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={320}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={320}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={320}
+            y={320}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={320}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={320}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={500}
+            y={320}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={0}
+            y={340}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={20}
+            y={340}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={40}
+            y={340}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={60}
+            y={340}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={340}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={340}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={340}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={340}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={340}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={420}
+            y={340}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={440}
+            y={340}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={540}
+            y={340}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={0}
+            y={360}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={60}
+            y={360}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={120}
+            y={360}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={140}
+            y={360}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={360}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={360}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={360}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={340}
+            y={360}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={380}
+            y={360}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={360}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={520}
+            y={360}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={540}
+            y={360}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={560}
+            y={360}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={0}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={320}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={380}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={420}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={440}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={460}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={540}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={560}
+            y={380}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={0}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={40}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={80}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={100}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={120}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={220}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={280}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={340}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={360}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={420}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={440}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={460}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={540}
+            y={400}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={420}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={420}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={420}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={280}
+            y={420}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={320}
+            y={420}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={360}
+            y={420}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={420}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={420}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={500}
+            y={420}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={540}
+            y={420}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={440}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={340}
+            y={440}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={380}
+            y={440}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={440}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={440}
+            y={440}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={440}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={500}
+            y={440}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={520}
+            y={440}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={460}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={460}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={220}
+            y={460}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={460}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={320}
+            y={460}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={340}
+            y={460}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={360}
+            y={460}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={460}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={460}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={500}
+            y={460}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={540}
+            y={460}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={560}
+            y={460}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={220}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={320}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={340}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={420}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={440}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={460}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={500}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={520}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={540}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={560}
+            y={480}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={500}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={500}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={220}
+            y={500}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={500}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={360}
+            y={500}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={400}
+            y={500}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={500}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={500}
+            y={500}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={520}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={520}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={520}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={220}
+            y={520}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={520}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={520}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={380}
+            y={520}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={460}
+            y={520}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={520}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={520}
+            y={520}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={540}
+            y={520}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={540}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={540}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={540}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={540}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={280}
+            y={540}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={320}
+            y={540}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={360}
+            y={540}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={420}
+            y={540}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={460}
+            y={540}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={520}
+            y={540}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={560}
+            y={540}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={160}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={180}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={200}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={220}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={240}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={260}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={300}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={320}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={340}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={360}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={420}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={440}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={460}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={500}
+            y={560}
+            width={20}
+            height={20}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <path
+            d="M0,0 h140 v140 h-140 Z M20.0,20.0 h100.0 v100.0 h-100.0 Z"
+            fillRule="evenodd"
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <path
+            d="M440,0 h140 v140 h-140 Z M460.0,20.0 h100.0 v100.0 h-100.0 Z"
+            fillRule="evenodd"
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={40}
+            y={40}
+            width={60}
+            height={60}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={480}
+            y={40}
+            width={60}
+            height={60}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <path
+            d="M0,440 h140 v140 h-140 Z M20.0,460.0 h100.0 v100.0 h-100.0 Z"
+            fillRule="evenodd"
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+          <rect
+            x={40}
+            y={480}
+            width={60}
+            height={60}
+            fill="rgb(0,0,0)"
+            opacity={1}
+            shapeRendering="crispEdges"
+          />
+        </svg>
+      </g>
+    </svg>
+  );
+}
