+++ packages/react-wallet-kit/src/providers/client/Provider.tsx
@@ -264,7 +264,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
   const [user, setUser] = useState<v1User | undefined>(undefined);
   const [clientState, setClientState] = useState<ClientState>();
   const [authState, setAuthState] = useState<AuthState>(
-    AuthState.Unauthenticated
+    AuthState.Unauthenticated,
   );
 
   // if there is no authProxyConfigId or if autoFetchWalletKitConfig is specifically
@@ -277,12 +277,12 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
   const [walletProviders, setWalletProviders] = useWalletProviderState();
 
   const [walletConnectApps, setWalletConnectApps] = useState<WalletProvider[]>(
-    []
+    [],
   );
 
   const expiryTimeoutsRef = useRef<TimerMap>({});
   const proxyAuthConfigRef = useRef<ProxyTGetWalletKitConfigResponse | null>(
-    null
+    null,
   );
 
   const [allSessions, setAllSessions] = useState<
@@ -709,7 +709,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
   };
 
   const buildConfig = (
-    proxyAuthConfig?: ProxyTGetWalletKitConfigResponse | undefined
+    proxyAuthConfig?: ProxyTGetWalletKitConfigResponse | undefined,
   ) => {
     // Juggle the local overrides with the values set in the dashboard (proxyAuthConfig).
     const resolvedMethods = {
@@ -788,7 +788,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
     const oauthOrder =
       config.ui?.authModal?.oauthOrder ??
       (["google", "apple", "x", "discord", "facebook"] as const).filter(
-        (provider) => resolvedMethods[`${provider}OauthEnabled` as const]
+        (provider) => resolvedMethods[`${provider}OauthEnabled` as const],
       );
 
     // Set a default ordering for the overall auth methods
@@ -809,17 +809,17 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
     if (proxyAuthConfig) {
       if (config.auth?.sessionExpirationSeconds) {
         console.warn(
-          "Turnkey SDK warning. You have set sessionExpirationSeconds directly in the TurnkeyProvider. This setting will be ignored because you are using an auth proxy. Please configure session expiration in the Turnkey dashboard."
+          "Turnkey SDK warning. You have set sessionExpirationSeconds directly in the TurnkeyProvider. This setting will be ignored because you are using an auth proxy. Please configure session expiration in the Turnkey dashboard.",
         );
       }
       if (config.auth?.otpAlphanumeric !== undefined) {
         console.warn(
-          "Turnkey SDK warning. You have set otpAlphanumeric directly in the TurnkeyProvider. This setting will be ignored because you are using an auth proxy. Please configure OTP settings in the Turnkey dashboard."
+          "Turnkey SDK warning. You have set otpAlphanumeric directly in the TurnkeyProvider. This setting will be ignored because you are using an auth proxy. Please configure OTP settings in the Turnkey dashboard.",
         );
       }
       if (config.auth?.otpLength) {
         console.warn(
-          "Turnkey SDK warning. You have set otpLength directly in the TurnkeyProvider. This setting will be ignored because you are using an auth proxy. Please configure OTP settings in the Turnkey dashboard."
+          "Turnkey SDK warning. You have set otpLength directly in the TurnkeyProvider. This setting will be ignored because you are using an auth proxy. Please configure OTP settings in the Turnkey dashboard.",
         );
       }
     }
@@ -954,8 +954,8 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           new TurnkeyError(
             `Failed to initialize Turnkey client`,
             TurnkeyErrorCodes.INITIALIZE_CLIENT_ERROR,
-            error
-          )
+            error,
+          ),
         );
       }
     }
@@ -988,7 +988,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
             sessionKey,
             expiry: session!.expiry,
           });
-        })
+        }),
       );
 
       setAllSessions(allLocalStorageSessions || undefined);
@@ -1038,8 +1038,8 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           new TurnkeyError(
             `Failed to initialize sessions`,
             TurnkeyErrorCodes.INITIALIZE_SESSION_ERROR,
-            error
-          )
+            error,
+          ),
         );
       }
     }
@@ -1067,7 +1067,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
    */
   async function initializeWalletProviderListeners(
     walletProviders: WalletProvider[],
-    onUpdateState: () => Promise<void>
+    onUpdateState: () => Promise<void>,
   ): Promise<() => void> {
     if (walletProviders.length === 0) return () => {};
 
@@ -1082,7 +1082,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           (provider) =>
             provider.chainInfo.namespace === Chain.Ethereum &&
             nativeOnly(provider) &&
-            provider.connectedAddresses.length > 0
+            provider.connectedAddresses.length > 0,
         )
       : [];
 
@@ -1091,7 +1091,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           (provider) =>
             provider.chainInfo.namespace === Chain.Solana &&
             nativeOnly(provider) &&
-            provider.connectedAddresses.length > 0
+            provider.connectedAddresses.length > 0,
         )
       : [];
 
@@ -1103,18 +1103,18 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
     // can detect proposal expiration events and display the new regenerated
     // URI for the UI
     const wcProviders = walletProviders.filter(
-      (p) => p.interfaceType === WalletInterfaceType.WalletConnect
+      (p) => p.interfaceType === WalletInterfaceType.WalletConnect,
     );
 
     // since all WalletConnect providers share the same underlying session
     // and emit identical events, we only attach listeners to a single provider
     const wcProvider = wcProviders.find(
-      (p) => p.interfaceType === WalletInterfaceType.WalletConnect
+      (p) => p.interfaceType === WalletInterfaceType.WalletConnect,
     );
 
     function attachEthereumListeners(
       provider: any,
-      onUpdateState: () => Promise<void>
+      onUpdateState: () => Promise<void>,
     ) {
       if (typeof provider.on !== "function") return;
 
@@ -1137,7 +1137,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
 
     function attachSolanaListeners(
       provider: any,
-      onUpdateState: () => Promise<void>
+      onUpdateState: () => Promise<void>,
     ) {
       const cleanups: Array<() => void> = [];
 
@@ -1155,7 +1155,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
     ethProviders.forEach((p) => {
       const cleanup = attachEthereumListeners(
         (p as any).provider,
-        onUpdateState
+        onUpdateState,
       );
       if (cleanup) cleanups.push(cleanup);
     });
@@ -1187,8 +1187,8 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
                 new TurnkeyError(
                   `WalletConnect initialization failed: ${evt.error || "Unknown error"}`,
                   TurnkeyErrorCodes.WALLET_CONNECT_INITIALIZATION_ERROR,
-                  evt.error
-                )
+                  evt.error,
+                ),
               );
 
               return;
@@ -1203,7 +1203,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
               const currentSession = await getSession();
               if (currentSession) {
                 const wcProviders = providers?.filter(
-                  (p) => p.interfaceType === WalletInterfaceType.WalletConnect
+                  (p) => p.interfaceType === WalletInterfaceType.WalletConnect,
                 );
 
                 const wcWallets = await fetchWallets({
@@ -1213,7 +1213,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
 
                 if (wcWallets.length > 0) {
                   setWallets((prev) =>
-                    mergeWalletsWithoutDuplicates(prev, wcWallets)
+                    mergeWalletsWithoutDuplicates(prev, wcWallets),
                   );
                 }
               }
@@ -1261,8 +1261,8 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           new TurnkeyError(
             "Failed to clear session timeouts",
             TurnkeyErrorCodes.CLEAR_SESSION_TIMEOUTS_ERROR,
-            error
-          )
+            error,
+          ),
         );
       }
     }
@@ -1344,8 +1344,8 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           new TurnkeyError(
             `Failed to handle post-authentication`,
             TurnkeyErrorCodes.HANDLE_POST_AUTH_ERROR,
-            error
-          )
+            error,
+          ),
         );
       }
     }
@@ -1403,12 +1403,12 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       return client.createHttpClient(params);
     },
-    [client]
+    [client],
   );
 
   const overrideApiKeyStamper = useCallback(
@@ -1443,13 +1443,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
     if (!client)
       throw new TurnkeyError(
         "Client is not initialized.",
-        TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+        TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
       );
     return withTurnkeyErrorHandling(
       () => client.getActiveSessionKey(),
       undefined,
       callbacks,
-      "Failed to get active session key"
+      "Failed to get active session key",
     );
   }, [client, callbacks]);
 
@@ -1458,7 +1458,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       await withTurnkeyErrorHandling(
@@ -1472,7 +1472,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         },
         undefined,
         callbacks,
-        "Failed to logout"
+        "Failed to logout",
       );
 
       return;
@@ -1485,16 +1485,16 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.getSession(params),
         undefined,
         callbacks,
-        "Failed to get session"
+        "Failed to get session",
       );
     },
-    [client, callbacks]
+    [client, callbacks],
   );
 
   const getAllSessions = useCallback(async (): Promise<
@@ -1503,13 +1503,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
     if (!client)
       throw new TurnkeyError(
         "Client is not initialized.",
-        TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+        TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
       );
     return withTurnkeyErrorHandling(
       () => client.getAllSessions(),
       undefined,
       callbacks,
-      "Failed to get all sessions"
+      "Failed to get all sessions",
     );
   }, [client, callbacks]);
 
@@ -1518,17 +1518,17 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       return withTurnkeyErrorHandling(
         () => client.createPasskey({ ...params }),
         undefined,
         callbacks,
-        "Failed to create passkey"
+        "Failed to create passkey",
       );
     },
-    [client, callbacks]
+    [client, callbacks],
   );
 
   const fetchUser = useCallback(
@@ -1536,16 +1536,16 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.fetchUser(params),
         () => logout(),
         callbacks,
-        "Failed to fetch user"
+        "Failed to fetch user",
       );
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const fetchWalletProviders = useCallback(
@@ -1553,7 +1553,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       return withTurnkeyErrorHandling(
@@ -1570,10 +1570,10 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         },
         undefined,
         callbacks,
-        "Failed to fetch wallet providers"
+        "Failed to fetch wallet providers",
       );
     },
-    [client, callbacks]
+    [client, callbacks],
   );
 
   const fetchWallets = useCallback(
@@ -1581,17 +1581,17 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       return withTurnkeyErrorHandling(
         () => client.fetchWallets(params),
         () => logout(),
         callbacks,
-        "Failed to fetch wallets"
+        "Failed to fetch wallets",
       );
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const refreshUser = useCallback(
@@ -1601,7 +1601,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
 
       const user = await withTurnkeyErrorHandling(
@@ -1613,7 +1613,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           }),
         () => logout(),
         callbacks,
-        "Failed to refresh user"
+        "Failed to refresh user",
       );
 
       if (user) {
@@ -1622,7 +1622,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
 
       return user;
     },
-    [client, callbacks, fetchUser, logout]
+    [client, callbacks, fetchUser, logout],
   );
 
   /**
@@ -1639,7 +1639,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!masterConfig?.autoRefreshManagedState) return undefined;
       return refreshUser(params);
     },
-    [masterConfig, refreshUser]
+    [masterConfig, refreshUser],
   );
 
   const refreshWallets = useCallback(
@@ -1649,14 +1649,14 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
 
       const walletProviders = await withTurnkeyErrorHandling(
         () => fetchWalletProviders(),
         undefined,
         callbacks,
-        "Failed to refresh wallets"
+        "Failed to refresh wallets",
       );
 
       const wallets = await withTurnkeyErrorHandling(
@@ -1669,7 +1669,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           }),
         () => logout(),
         callbacks,
-        "Failed to refresh wallets"
+        "Failed to refresh wallets",
       );
 
       if (wallets) {
@@ -1678,7 +1678,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
 
       return wallets;
     },
-    [client, callbacks, fetchWalletProviders, fetchWallets, logout]
+    [client, callbacks, fetchWalletProviders, fetchWallets, logout],
   );
 
   /**
@@ -1695,24 +1695,23 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!masterConfig?.autoRefreshManagedState) return [];
       return refreshWallets(params);
     },
-    [masterConfig, refreshWallets]
+    [masterConfig, refreshWallets],
   );
 
-  // TODO (Amir): Comment and make comment header in method-types
   const fetchAndBuildWalletConnectApps = useCallback(async (): Promise<
     WalletProvider[]
   > => {
     if (!client) {
       throw new TurnkeyError(
         "Client is not initialized.",
-        TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+        TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
       );
     }
 
     if (!masterConfig?.walletConfig?.walletConnect?.projectId) {
       throw new TurnkeyError(
         "WalletConnect project ID is not configured.",
-        TurnkeyErrorCodes.WALLET_CONNECT_INITIALIZATION_ERROR
+        TurnkeyErrorCodes.WALLET_CONNECT_INITIALIZATION_ERROR,
       );
     }
 
@@ -1724,7 +1723,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       },
       undefined,
       callbacks,
-      "Failed to fetch and build WalletConnect apps"
+      "Failed to fetch and build WalletConnect apps",
     );
 
     setWalletConnectApps(providers);
@@ -1736,14 +1735,14 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const activeSessionKey = await getActiveSessionKey();
       await withTurnkeyErrorHandling(
         async () => client.clearSession(params),
         undefined,
         callbacks,
-        "Failed to clear session"
+        "Failed to clear session",
       );
       const sessionKey = params?.sessionKey ?? activeSessionKey;
       if (!sessionKey) return;
@@ -1810,7 +1809,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
               expiryTimeoutsRef.current,
               warnKey,
               runRefresh,
-              10_000
+              10_000,
             );
             return;
           }
@@ -1869,7 +1868,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
             expiryTimeoutsRef.current,
             warnKey,
             runRefresh,
-            warnAt - now
+            warnAt - now,
           );
         }
 
@@ -1878,7 +1877,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           expiryTimeoutsRef.current,
           sessionKey,
           expireSession,
-          timeUntilExpiry
+          timeUntilExpiry,
         );
       } catch (error) {
         if (
@@ -1891,8 +1890,8 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
             new TurnkeyError(
               `Failed to schedule session expiration for ${sessionKey}`,
               TurnkeyErrorCodes.SCHEDULE_SESSION_EXPIRY_ERROR,
-              error
-            )
+              error,
+            ),
           );
         }
       }
@@ -1910,17 +1909,17 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       getActiveSessionKey,
       clearSession,
       logout,
-    ]
+    ],
   );
 
   const refreshSession = useCallback(
     async (
-      params?: RefreshSessionParams
+      params?: RefreshSessionParams,
     ): Promise<TStampLoginResponse | undefined> => {
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
 
@@ -1928,7 +1927,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!activeSessionKey) {
         throw new TurnkeyError(
           "No active session found.",
-          TurnkeyErrorCodes.NO_SESSION_FOUND
+          TurnkeyErrorCodes.NO_SESSION_FOUND,
         );
       }
 
@@ -1938,7 +1937,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         () => client.refreshSession({ ...params }),
         () => logout(),
         callbacks,
-        "Failed to refresh session"
+        "Failed to refresh session",
       );
       const session = await getSession({ sessionKey });
 
@@ -1961,7 +1960,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       getSession,
       scheduleSessionExpiration,
       getAllSessions,
-    ]
+    ],
   );
 
   const loginWithPasskey = useCallback(
@@ -1969,7 +1968,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
 
@@ -1981,7 +1980,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         () => client.loginWithPasskey({ ...params, expirationSeconds }),
         () => logout(),
         callbacks,
-        "Failed to login with passkey"
+        "Failed to login with passkey",
       );
       if (res) {
         await handlePostAuth({
@@ -1993,7 +1992,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       }
       return res;
     },
-    [client, callbacks, masterConfig, logout, handlePostAuth]
+    [client, callbacks, masterConfig, logout, handlePostAuth],
   );
 
   const signUpWithPasskey = useCallback(
@@ -2001,13 +2000,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       if (!masterConfig) {
         throw new TurnkeyError(
           "Config is not ready yet!",
-          TurnkeyErrorCodes.INVALID_CONFIGURATION
+          TurnkeyErrorCodes.INVALID_CONFIGURATION,
         );
       }
       // If createSubOrgParams is not provided, use the default from masterConfig
@@ -2050,7 +2049,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           }),
         () => logout(),
         callbacks,
-        "Failed to sign up with passkey"
+        "Failed to sign up with passkey",
       );
       if (res) {
         await handlePostAuth({
@@ -2062,7 +2061,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       }
       return res;
     },
-    [client, callbacks, logout, handlePostAuth, masterConfig]
+    [client, callbacks, logout, handlePostAuth, masterConfig],
   );
 
   const connectWalletAccount = useCallback(
@@ -2070,7 +2069,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
 
@@ -2091,7 +2090,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           // we narrow to only connected wallets
           // because we know the account must come from one of them
           const connectedWallets = wallets.filter(
-            (w): w is ConnectedWallet => w.source === WalletSource.Connected
+            (w): w is ConnectedWallet => w.source === WalletSource.Connected,
           );
 
           // find the matching account
@@ -2102,7 +2101,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           if (!matchedAccount) {
             throw new TurnkeyError(
               `No connected wallet account found for address: ${address}`,
-              TurnkeyErrorCodes.NO_WALLET_FOUND
+              TurnkeyErrorCodes.NO_WALLET_FOUND,
             );
           }
 
@@ -2110,11 +2109,11 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         },
         () => logout(),
         callbacks,
-        "Failed to connect wallet account"
+        "Failed to connect wallet account",
       );
     },
 
-    [client, callbacks, getSession, logout, maybeRefreshWallets, fetchWallets]
+    [client, callbacks, getSession, logout, maybeRefreshWallets, fetchWallets],
   );
 
   const disconnectWalletAccount = useCallback(
@@ -2122,7 +2121,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
 
@@ -2135,10 +2134,10 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         },
         undefined,
         callbacks,
-        "Failed to disconnect wallet account"
+        "Failed to disconnect wallet account",
       );
     },
-    [client, callbacks]
+    [client, callbacks],
   );
 
   const switchWalletAccountChain = useCallback(
@@ -2146,7 +2145,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
 
@@ -2156,20 +2155,20 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         },
         undefined,
         callbacks,
-        "Failed to switch wallet account chain"
+        "Failed to switch wallet account chain",
       );
     },
-    [client, walletProviders, callbacks]
+    [client, walletProviders, callbacks],
   );
 
   const buildWalletLoginRequest = useCallback(
     async (
-      params: BuildWalletLoginRequestParams
+      params: BuildWalletLoginRequestParams,
     ): Promise<BuildWalletLoginRequestResult> => {
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
 
@@ -2181,10 +2180,10 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         () => client.buildWalletLoginRequest({ ...params, expirationSeconds }),
         undefined,
         callbacks,
-        "Failed to login with wallet"
+        "Failed to login with wallet",
       );
     },
-    [client, callbacks, masterConfig]
+    [client, callbacks, masterConfig],
   );
 
   const loginWithWallet = useCallback(
@@ -2192,7 +2191,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
 
@@ -2204,7 +2203,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         () => client.loginWithWallet({ ...params, expirationSeconds }),
         undefined,
         callbacks,
-        "Failed to login with wallet"
+        "Failed to login with wallet",
       );
       if (res) {
         await handlePostAuth({
@@ -2216,7 +2215,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       }
       return res;
     },
-    [client, callbacks, handlePostAuth, masterConfig]
+    [client, callbacks, handlePostAuth, masterConfig],
   );
 
   const signUpWithWallet = useCallback(
@@ -2224,13 +2223,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       if (!masterConfig) {
         throw new TurnkeyError(
           "Config is not ready yet!",
-          TurnkeyErrorCodes.INVALID_CONFIGURATION
+          TurnkeyErrorCodes.INVALID_CONFIGURATION,
         );
       }
       // If createSubOrgParams is not provided, use the default from masterConfig
@@ -2250,7 +2249,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         () => client.signUpWithWallet({ ...params, expirationSeconds }),
         undefined,
         callbacks,
-        "Failed to sign up with wallet"
+        "Failed to sign up with wallet",
       );
       if (res) {
         await handlePostAuth({
@@ -2262,23 +2261,23 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       }
       return res;
     },
-    [client, callbacks, handlePostAuth, masterConfig]
+    [client, callbacks, handlePostAuth, masterConfig],
   );
 
   const loginOrSignupWithWallet = useCallback(
     async (
-      params: LoginOrSignupWithWalletParams
+      params: LoginOrSignupWithWalletParams,
     ): Promise<WalletAuthResult & { action: AuthAction }> => {
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       if (!masterConfig) {
         throw new TurnkeyError(
           "Config is not ready yet!",
-          TurnkeyErrorCodes.INVALID_CONFIGURATION
+          TurnkeyErrorCodes.INVALID_CONFIGURATION,
         );
       }
       // If createSubOrgParams is not provided, use the default from masterConfig
@@ -2298,7 +2297,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         () => client.loginOrSignupWithWallet({ ...params, expirationSeconds }),
         undefined,
         callbacks,
-        "Failed to login or sign up with wallet"
+        "Failed to login or sign up with wallet",
       );
       if (res) {
         await handlePostAuth({
@@ -2310,7 +2309,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       }
       return res;
     },
-    [client, callbacks, handlePostAuth, masterConfig]
+    [client, callbacks, handlePostAuth, masterConfig],
   );
 
   const initOtp = useCallback(
@@ -2318,17 +2317,17 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       return withTurnkeyErrorHandling(
         () => client.initOtp(params),
         undefined,
         callbacks,
-        "Failed to initialize OTP"
+        "Failed to initialize OTP",
       );
     },
-    [client, callbacks]
+    [client, callbacks],
   );
 
   const verifyOtp = useCallback(
@@ -2336,17 +2335,17 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       return withTurnkeyErrorHandling(
         () => client.verifyOtp(params),
         undefined,
         callbacks,
-        "Failed to verify OTP"
+        "Failed to verify OTP",
       );
     },
-    [client, callbacks]
+    [client, callbacks],
   );
 
   const loginWithOtp = useCallback(
@@ -2354,7 +2353,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
 
@@ -2362,7 +2361,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         () => client.loginWithOtp(params),
         undefined,
         callbacks,
-        "Failed to login with OTP"
+        "Failed to login with OTP",
       );
       if (res) {
         await handlePostAuth({
@@ -2374,7 +2373,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       }
       return res;
     },
-    [client, callbacks, handlePostAuth]
+    [client, callbacks, handlePostAuth],
   );
 
   const signUpWithOtp = useCallback(
@@ -2382,13 +2381,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       if (!masterConfig) {
         throw new TurnkeyError(
           "Config is not ready yet!",
-          TurnkeyErrorCodes.INVALID_CONFIGURATION
+          TurnkeyErrorCodes.INVALID_CONFIGURATION,
         );
       }
       // If createSubOrgParams is not provided, use the default from masterConfig
@@ -2410,7 +2409,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         () => client.signUpWithOtp(params),
         undefined,
         callbacks,
-        "Failed to sign up with OTP"
+        "Failed to sign up with OTP",
       );
       if (res) {
         await handlePostAuth({
@@ -2422,25 +2421,25 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       }
       return res;
     },
-    [client, callbacks, masterConfig, handlePostAuth]
+    [client, callbacks, masterConfig, handlePostAuth],
   );
 
   const completeOtp = useCallback(
     async (
-      params: CompleteOtpParams
+      params: CompleteOtpParams,
     ): Promise<
       BaseAuthResult & { verificationToken: string; action: AuthAction }
     > => {
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       if (!masterConfig) {
         throw new TurnkeyError(
           "Config is not ready yet!",
-          TurnkeyErrorCodes.INVALID_CONFIGURATION
+          TurnkeyErrorCodes.INVALID_CONFIGURATION,
         );
       }
 
@@ -2463,7 +2462,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         () => client.completeOtp(params),
         undefined,
         callbacks,
-        "Failed to complete OTP"
+        "Failed to complete OTP",
       );
       if (res) {
         await handlePostAuth({
@@ -2475,7 +2474,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       }
       return res;
     },
-    [client, callbacks, handlePostAuth, masterConfig]
+    [client, callbacks, handlePostAuth, masterConfig],
   );
 
   const loginWithOauth = useCallback(
@@ -2483,7 +2482,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
 
@@ -2491,7 +2490,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         () => client.loginWithOauth(params),
         undefined,
         callbacks,
-        "Failed to login with OAuth"
+        "Failed to login with OAuth",
       );
       if (res) {
         await handlePostAuth({
@@ -2503,7 +2502,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       }
       return res;
     },
-    [client, callbacks, handlePostAuth]
+    [client, callbacks, handlePostAuth],
   );
 
   const signUpWithOauth = useCallback(
@@ -2511,13 +2510,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       if (!masterConfig) {
         throw new TurnkeyError(
           "Config is not ready yet!",
-          TurnkeyErrorCodes.INVALID_CONFIGURATION
+          TurnkeyErrorCodes.INVALID_CONFIGURATION,
         );
       }
       // If createSubOrgParams is not provided, use the default from masterConfig
@@ -2533,7 +2532,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         () => client.signUpWithOauth(params),
         undefined,
         callbacks,
-        "Failed to sign up with OAuth"
+        "Failed to sign up with OAuth",
       );
       if (res) {
         await handlePostAuth({
@@ -2545,23 +2544,23 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       }
       return res;
     },
-    [client, callbacks, handlePostAuth, masterConfig]
+    [client, callbacks, handlePostAuth, masterConfig],
   );
 
   const completeOauth = useCallback(
     async (
-      params: CompleteOauthParams
+      params: CompleteOauthParams,
     ): Promise<BaseAuthResult & { action: AuthAction }> => {
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       if (!masterConfig) {
         throw new TurnkeyError(
           "Config is not ready yet!",
-          TurnkeyErrorCodes.INVALID_CONFIGURATION
+          TurnkeyErrorCodes.INVALID_CONFIGURATION,
         );
       }
 
@@ -2579,7 +2578,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         () => client.completeOauth(params),
         undefined,
         callbacks,
-        "Failed to complete OAuth"
+        "Failed to complete OAuth",
       );
       if (res) {
         await handlePostAuth({
@@ -2591,7 +2590,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       }
       return res;
     },
-    [client, callbacks, masterConfig, handlePostAuth]
+    [client, callbacks, masterConfig, handlePostAuth],
   );
 
   const fetchWalletAccounts = useCallback(
@@ -2599,17 +2598,17 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       return withTurnkeyErrorHandling(
         () => client.fetchWalletAccounts(params),
         () => logout(),
         callbacks,
-        "Failed to fetch wallet accounts"
+        "Failed to fetch wallet accounts",
       );
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const fetchPrivateKeys = useCallback(
@@ -2617,17 +2616,17 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
       return withTurnkeyErrorHandling(
         () => client.fetchPrivateKeys(params),
         () => logout(),
         callbacks,
-        "Failed to fetch private keys"
+        "Failed to fetch private keys",
       );
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const signMessage = useCallback(
@@ -2635,27 +2634,27 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.signMessage(params),
         () => logout(),
         callbacks,
-        "Failed to sign message"
+        "Failed to sign message",
       );
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const handleSignMessage = useCallback(
     async (
-      params: HandleSignMessageParams
+      params: HandleSignMessageParams,
     ): Promise<v1SignRawPayloadResult> => {
       const { successPageDuration = 2000 } = params;
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
 
       return new Promise((resolve, reject) => {
@@ -2690,13 +2689,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
             reject(
               new TurnkeyError(
                 "User canceled the signing process.",
-                TurnkeyErrorCodes.USER_CANCELED
-              )
+                TurnkeyErrorCodes.USER_CANCELED,
+              ),
             ),
         });
       });
     },
-    [client, callbacks, pushPage]
+    [client, callbacks, pushPage],
   );
 
   const signTransaction = useCallback(
@@ -2704,16 +2703,16 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.signTransaction(params),
         () => logout(),
         callbacks,
-        "Failed to sign transaction"
+        "Failed to sign transaction",
       );
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const ethSendTransaction = useCallback(
@@ -2721,16 +2720,16 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.ethSendTransaction(params),
         () => logout(),
         callbacks,
-        "Failed to send eth transaction"
+        "Failed to send eth transaction",
       );
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const ethSendErc20Transfer = useCallback(
@@ -2772,34 +2771,34 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.signAndSendTransaction(params),
         () => logout(),
         callbacks,
-        "Failed to sign and send transaction"
+        "Failed to sign and send transaction",
       );
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
   const pollTransactionStatus = useCallback(
     async (
-      params: PollTransactionStatusParams
+      params: PollTransactionStatusParams,
     ): Promise<TGetSendTransactionStatusResponse> => {
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.pollTransactionStatus(params),
         () => logout(),
         callbacks,
-        "Failed to poll transaction status"
+        "Failed to poll transaction status",
       );
     },
-    [client, callbacks]
+    [client, callbacks],
   );
 
   const fetchOrCreateP256ApiKeyUser = useCallback(
@@ -2807,35 +2806,35 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.fetchOrCreateP256ApiKeyUser(params),
         () => logout(),
         callbacks,
-        "Failed to fetch or create delegated access user"
+        "Failed to fetch or create delegated access user",
       );
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const fetchOrCreatePolicies = useCallback(
     async (
-      params: FetchOrCreatePoliciesParams
+      params: FetchOrCreatePoliciesParams,
     ): Promise<FetchOrCreatePoliciesResult> => {
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.fetchOrCreatePolicies(params),
         () => logout(),
         callbacks,
-        "Failed to fetch or create delegated access user"
+        "Failed to fetch or create delegated access user",
       );
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const updateUserEmail = useCallback(
@@ -2843,13 +2842,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.updateUserEmail(params),
         () => logout(),
         callbacks,
-        "Failed to update user email"
+        "Failed to update user email",
       );
       if (res)
         await maybeRefreshUser({
@@ -2861,7 +2860,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         });
       return res;
     },
-    [client, callbacks, logout, maybeRefreshUser]
+    [client, callbacks, logout, maybeRefreshUser],
   );
 
   const removeUserEmail = useCallback(
@@ -2869,13 +2868,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.removeUserEmail(params),
         () => logout(),
         callbacks,
-        "Failed to remove user email"
+        "Failed to remove user email",
       );
       if (res)
         await maybeRefreshUser({
@@ -2887,7 +2886,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         });
       return res;
     },
-    [client, callbacks, logout, maybeRefreshUser]
+    [client, callbacks, logout, maybeRefreshUser],
   );
 
   const updateUserPhoneNumber = useCallback(
@@ -2895,13 +2894,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.updateUserPhoneNumber(params),
         () => logout(),
         callbacks,
-        "Failed to update user phone number"
+        "Failed to update user phone number",
       );
       if (res)
         await maybeRefreshUser({
@@ -2913,7 +2912,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         });
       return res;
     },
-    [client, callbacks, logout, maybeRefreshUser]
+    [client, callbacks, logout, maybeRefreshUser],
   );
 
   const removeUserPhoneNumber = useCallback(
@@ -2921,13 +2920,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.removeUserPhoneNumber(params),
         () => logout(),
         callbacks,
-        "Failed to remove user phone number"
+        "Failed to remove user phone number",
       );
       if (res)
         await maybeRefreshUser({
@@ -2939,7 +2938,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         });
       return res;
     },
-    [client, callbacks, logout, maybeRefreshUser]
+    [client, callbacks, logout, maybeRefreshUser],
   );
 
   const updateUserName = useCallback(
@@ -2947,13 +2946,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.updateUserName(params),
         () => logout(),
         callbacks,
-        "Failed to update user name"
+        "Failed to update user name",
       );
       if (res)
         await maybeRefreshUser({
@@ -2965,7 +2964,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         });
       return res;
     },
-    [client, callbacks, logout, maybeRefreshUser]
+    [client, callbacks, logout, maybeRefreshUser],
   );
 
   const addOauthProvider = useCallback(
@@ -2973,13 +2972,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.addOauthProvider(params),
         () => logout(),
         callbacks,
-        "Failed to add OAuth provider"
+        "Failed to add OAuth provider",
       );
       if (res)
         await maybeRefreshUser({
@@ -2991,7 +2990,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         });
       return res;
     },
-    [client, callbacks, logout, maybeRefreshUser]
+    [client, callbacks, logout, maybeRefreshUser],
   );
 
   const removeOauthProviders = useCallback(
@@ -2999,13 +2998,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.removeOauthProviders(params),
         () => logout(),
         callbacks,
-        "Failed to remove OAuth providers"
+        "Failed to remove OAuth providers",
       );
       if (res)
         await maybeRefreshUser({
@@ -3017,7 +3016,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         });
       return res;
     },
-    [client, callbacks, logout, maybeRefreshUser]
+    [client, callbacks, logout, maybeRefreshUser],
   );
 
   const addPasskey = useCallback(
@@ -3025,13 +3024,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.addPasskey(params),
         () => logout(),
         callbacks,
-        "Failed to add passkey"
+        "Failed to add passkey",
       );
       if (res) {
         const updatedUser = await maybeRefreshUser({
@@ -3065,13 +3064,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.removePasskeys(params),
         () => logout(),
         callbacks,
-        "Failed to remove passkeys"
+        "Failed to remove passkeys",
       );
       if (res)
         await maybeRefreshUser({
@@ -3083,7 +3082,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         });
       return res;
     },
-    [client, callbacks, logout, maybeRefreshUser]
+    [client, callbacks, logout, maybeRefreshUser],
   );
 
   const createWallet = useCallback(
@@ -3091,13 +3090,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.createWallet(params),
         () => logout(),
         callbacks,
-        "Failed to create wallet"
+        "Failed to create wallet",
       );
       const s = await getSession();
       if (res && s)
@@ -3109,7 +3108,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         });
       return res;
     },
-    [client, session, callbacks, logout, maybeRefreshWallets, getSession]
+    [client, session, callbacks, logout, maybeRefreshWallets, getSession],
   );
 
   const createWalletAccounts = useCallback(
@@ -3117,13 +3116,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.createWalletAccounts(params),
         () => logout(),
         callbacks,
-        "Failed to create wallet accounts"
+        "Failed to create wallet accounts",
       );
       const s = await getSession();
       if (res && s)
@@ -3135,7 +3134,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         });
       return res;
     },
-    [client, session, callbacks, logout, getSession, maybeRefreshWallets]
+    [client, session, callbacks, logout, getSession, maybeRefreshWallets],
   );
 
   const exportWallet = useCallback(
@@ -3143,13 +3142,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.exportWallet(params),
         () => logout(),
         callbacks,
-        "Failed to export wallet"
+        "Failed to export wallet",
       );
       const s = await getSession();
       if (res && s)
@@ -3161,7 +3160,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         });
       return res;
     },
-    [client, session, callbacks, logout, getSession, maybeRefreshWallets]
+    [client, session, callbacks, logout, getSession, maybeRefreshWallets],
   );
 
   const exportPrivateKey = useCallback(
@@ -3169,17 +3168,17 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.exportPrivateKey(params),
         () => logout(),
         callbacks,
-        "Failed to export private key"
+        "Failed to export private key",
       );
       return res;
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const exportWalletAccount = useCallback(
@@ -3187,13 +3186,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.exportWalletAccount(params),
         () => logout(),
         callbacks,
-        "Failed to export wallet accounts"
+        "Failed to export wallet accounts",
       );
       const s = await getSession();
       if (res && s)
@@ -3205,7 +3204,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         });
       return res;
     },
-    [client, callbacks, logout, getSession, maybeRefreshWallets]
+    [client, callbacks, logout, getSession, maybeRefreshWallets],
   );
 
   const importWallet = useCallback(
@@ -3213,13 +3212,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.importWallet(params),
         () => logout(),
         callbacks,
-        "Failed to import wallet"
+        "Failed to import wallet",
       );
       const s = await getSession();
       if (res && s)
@@ -3232,7 +3231,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         });
       return res;
     },
-    [client, callbacks, logout, getSession, maybeRefreshWallets]
+    [client, callbacks, logout, getSession, maybeRefreshWallets],
   );
 
   const importPrivateKey = useCallback(
@@ -3240,36 +3239,36 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const res = await withTurnkeyErrorHandling(
         () => client.importPrivateKey(params),
         () => logout(),
         callbacks,
-        "Failed to import private key"
+        "Failed to import private key",
       );
       return res;
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const deleteSubOrganization = useCallback(
     async (
-      params?: DeleteSubOrganizationParams
+      params?: DeleteSubOrganizationParams,
     ): Promise<TDeleteSubOrganizationResponse> => {
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.deleteSubOrganization(params),
         () => logout(),
         callbacks,
-        "Failed to delete sub-organization"
+        "Failed to delete sub-organization",
       );
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const storeSession = useCallback(
@@ -3277,13 +3276,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       await withTurnkeyErrorHandling(
         () => client.storeSession(params),
         () => logout(),
         callbacks,
-        "Failed to store session"
+        "Failed to store session",
       );
       const sessionKey = await getActiveSessionKey();
       const session = await getSession({
@@ -3320,14 +3319,14 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       masterConfig?.auth?.scopePasskeyToUser,
       maybeRefreshWallets,
       maybeRefreshUser,
-    ]
+    ],
   );
 
   const clearAllSessions = useCallback(async (): Promise<void> => {
     if (!client)
       throw new TurnkeyError(
         "Client is not initialized.",
-        TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+        TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
       );
     setSession(undefined);
     setAllSessions(undefined);
@@ -3338,7 +3337,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       () => client.clearAllSessions(),
       () => logout(),
       callbacks,
-      "Failed to clear all sessions"
+      "Failed to clear all sessions",
     );
     await resetPasskeyScope(client, client.config.passkeyConfig);
   }, [
@@ -3354,26 +3353,26 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const session = await withTurnkeyErrorHandling(
         () => client.getSession({ sessionKey: params.sessionKey }),
         () => logout(),
         callbacks,
-        "Failed to get session"
+        "Failed to get session",
       );
       const s = await getSession();
       if (!s) {
         throw new TurnkeyError(
           "Session not found.",
-          TurnkeyErrorCodes.NOT_FOUND
+          TurnkeyErrorCodes.NOT_FOUND,
         );
       }
       await withTurnkeyErrorHandling(
         () => client.setActiveSession(params),
         () => logout(),
         callbacks,
-        "Failed to set active session"
+        "Failed to set active session",
       );
       setSession(session);
       await withTurnkeyErrorHandling(
@@ -3390,7 +3389,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         },
         () => logout(),
         callbacks,
-        "Failed to refresh data after setting active session"
+        "Failed to refresh data after setting active session",
       );
       return;
     },
@@ -3403,20 +3402,20 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       masterConfig?.auth?.scopePasskeyToUser,
       maybeRefreshWallets,
       maybeRefreshUser,
-    ]
+    ],
   );
 
   const clearUnusedKeyPairs = useCallback(async (): Promise<void> => {
     if (!client)
       throw new TurnkeyError(
         "Client is not initialized.",
-        TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+        TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
       );
     return withTurnkeyErrorHandling(
       () => client.clearUnusedKeyPairs(),
       undefined,
       callbacks,
-      "Failed to clear unused key pairs"
+      "Failed to clear unused key pairs",
     );
   }, [client, callbacks]);
 
@@ -3425,16 +3424,16 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.createApiKeyPair(params),
         undefined,
         callbacks,
-        "Failed to create API key pair"
+        "Failed to create API key pair",
       );
     },
-    [client, callbacks]
+    [client, callbacks],
   );
 
   const deleteApiKeyPair = useCallback(
@@ -3476,13 +3475,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.getProxyAuthConfig(),
         undefined,
         callbacks,
-        "Failed to get proxy auth config"
+        "Failed to get proxy auth config",
       );
     }, [client, callbacks]);
 
@@ -3491,16 +3490,16 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.fetchBootProofForAppProof(params),
         () => logout(),
         callbacks,
-        "Failed to fetch or create delegated access user"
+        "Failed to fetch or create delegated access user",
       );
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const verifyAppProofs = useCallback(
@@ -3508,16 +3507,16 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       return withTurnkeyErrorHandling(
         () => client.verifyAppProofs(params),
         () => logout(),
         callbacks,
-        "Failed to verify app proofs"
+        "Failed to verify app proofs",
       );
     },
-    [client, callbacks, logout]
+    [client, callbacks, logout],
   );
 
   const handleDiscordOauth = useCallback(
@@ -3664,7 +3663,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         }
       });
     },
-    [client, callbacks, completeOauth, createApiKeyPair, masterConfig]
+    [client, callbacks, completeOauth, createApiKeyPair, masterConfig],
   );
 
   const handleXOauth = useCallback(
@@ -3810,7 +3809,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         }
       });
     },
-    [client, callbacks, completeOauth, createApiKeyPair, masterConfig]
+    [client, callbacks, completeOauth, createApiKeyPair, masterConfig],
   );
 
   const handleGoogleOauth = useCallback(
@@ -3942,7 +3941,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         }
       });
     },
-    [callbacks, completeOauth, createApiKeyPair, masterConfig]
+    [callbacks, completeOauth, createApiKeyPair, masterConfig],
   );
 
   const handleAppleOauth = useCallback(
@@ -4073,7 +4072,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         }
       });
     },
-    [callbacks, completeOauth, createApiKeyPair, masterConfig]
+    [callbacks, completeOauth, createApiKeyPair, masterConfig],
   );
 
   const handleFacebookOauth = useCallback(
@@ -4217,7 +4216,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         }
       });
     },
-    [callbacks, completeOauth, createApiKeyPair, masterConfig]
+    [callbacks, completeOauth, createApiKeyPair, masterConfig],
   );
 
   const handleLogin = useCallback(
@@ -4238,7 +4237,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         showTitle: logo ? false : true,
       });
     },
-    [pushPage, masterConfig]
+    [pushPage, masterConfig],
   );
 
   const handleExportWallet = useCallback(
@@ -4263,13 +4262,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
             reject(
               new TurnkeyError(
                 "User canceled the export wallet process.",
-                TurnkeyErrorCodes.USER_CANCELED
-              )
+                TurnkeyErrorCodes.USER_CANCELED,
+              ),
             ),
-        })
+        }),
       );
     },
-    [pushPage]
+    [pushPage],
   );
 
   const handleExportPrivateKey = useCallback(
@@ -4300,13 +4299,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
             reject(
               new TurnkeyError(
                 "User canceled the export private key process.",
-                TurnkeyErrorCodes.USER_CANCELED
-              )
+                TurnkeyErrorCodes.USER_CANCELED,
+              ),
             ),
-        })
+        }),
       );
     },
-    [pushPage]
+    [pushPage],
   );
 
   const handleExportWalletAccount = useCallback(
@@ -4333,13 +4332,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
             reject(
               new TurnkeyError(
                 "User canceled the export wallet account process.",
-                TurnkeyErrorCodes.USER_CANCELED
-              )
+                TurnkeyErrorCodes.USER_CANCELED,
+              ),
             ),
-        })
+        }),
       );
     },
-    [pushPage]
+    [pushPage],
   );
 
   const handleImportWallet = useCallback(
@@ -4381,9 +4380,9 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
                     {...(userId !== undefined && { userId })}
                   />
                 ),
-              })
+              }),
             ),
-          () => logout()
+          () => logout(),
         );
       } catch (error) {
         if (error instanceof TurnkeyError) {
@@ -4392,11 +4391,11 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         throw new TurnkeyError(
           "Failed to import wallet.",
           TurnkeyErrorCodes.IMPORT_WALLET_ERROR,
-          error
+          error,
         );
       }
     },
-    [pushPage, logout]
+    [pushPage, logout],
   );
 
   const handleImportPrivateKey = useCallback(
@@ -4438,9 +4437,9 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
                     {...(organizationId !== undefined && { organizationId })}
                   />
                 ),
-              })
+              }),
             ),
-          () => logout()
+          () => logout(),
         );
       } catch (error) {
         if (error instanceof TurnkeyError) {
@@ -4449,11 +4448,11 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         throw new TurnkeyError(
           "Failed to import private key.",
           TurnkeyErrorCodes.IMPORT_WALLET_ERROR,
-          error
+          error,
         );
       }
     },
-    [pushPage, logout]
+    [pushPage, logout],
   );
 
   const handleUpdateUserName = useCallback(
@@ -4468,7 +4467,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
 
       const s = await getSession();
@@ -4476,14 +4475,14 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
       const userId = params?.userId || user?.userId;
       if (!userId) {
         throw new TurnkeyError(
           "A user ID is required to update a user name.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
 
@@ -4531,7 +4530,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
                   showTitle: false,
                 });
               }),
-            () => logout()
+            () => logout(),
           );
         } else {
           const res = await updateUserName({
@@ -4550,11 +4549,11 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         throw new TurnkeyError(
           "Failed to update user name.",
           TurnkeyErrorCodes.UPDATE_USER_NAME_ERROR,
-          error
+          error,
         );
       }
     },
-    [client, getSession, user, pushPage, closeModal, updateUserName, logout]
+    [client, getSession, user, pushPage, closeModal, updateUserName, logout],
   );
 
   const handleUpdateUserPhoneNumber = useCallback(
@@ -4569,13 +4568,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
 
       if (!masterConfig) {
         throw new TurnkeyError(
           "Config is not ready yet!",
-          TurnkeyErrorCodes.CONFIG_NOT_INITIALIZED
+          TurnkeyErrorCodes.CONFIG_NOT_INITIALIZED,
         );
       }
 
@@ -4584,14 +4583,14 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
       const userId = params?.userId || s?.userId;
       if (!userId) {
         throw new TurnkeyError(
           "A user ID is required to update a phone number.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
 
@@ -4635,7 +4634,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
                   showTitle: false,
                 });
               }),
-            () => logout()
+            () => logout(),
           );
         } else {
           const { otpId, otpEncryptionTargetBundle } = await initOtp({
@@ -4693,7 +4692,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
               }),
             () => logout(),
             callbacks,
-            "Failed to update phone number"
+            "Failed to update phone number",
           );
         }
       } catch (error) {
@@ -4703,7 +4702,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         throw new TurnkeyError(
           "Failed to initialize OTP for sms verification.",
           TurnkeyErrorCodes.INIT_OTP_ERROR,
-          error
+          error,
         );
       }
     },
@@ -4719,7 +4718,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       updateUserPhoneNumber,
       user,
       callbacks,
-    ]
+    ],
   );
 
   const handleUpdateUserEmail = useCallback(
@@ -4734,7 +4733,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
 
       const s = await getSession();
@@ -4742,14 +4741,14 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
       const userId = params?.userId || s?.userId;
       if (!userId) {
         throw new TurnkeyError(
           "A user ID is required to update an email.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
 
@@ -4795,7 +4794,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
                   showTitle: false,
                 });
               }),
-            () => logout()
+            () => logout(),
           );
         } else {
           const { otpId, otpEncryptionTargetBundle } = await initOtp({
@@ -4850,7 +4849,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
               }),
             () => logout(),
             callbacks,
-            "Failed to update email"
+            "Failed to update email",
           );
         }
       } catch (error) {
@@ -4860,7 +4859,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         throw new TurnkeyError(
           "Failed to initialize OTP for email verification.",
           TurnkeyErrorCodes.INIT_OTP_ERROR,
-          error
+          error,
         );
       }
     },
@@ -4876,7 +4875,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       updateUserEmail,
       user,
       callbacks,
-    ]
+    ],
   );
 
   const handleAddEmail = useCallback(
@@ -4890,7 +4889,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client) {
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       }
 
@@ -4899,14 +4898,14 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
       const userId = params?.userId || s?.userId;
       if (!userId) {
         throw new TurnkeyError(
           "A user ID is required to add an email.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
 
@@ -4954,7 +4953,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
                   showTitle: false,
                 });
               }),
-            () => logout()
+            () => logout(),
           );
         } else {
           const { otpId, otpEncryptionTargetBundle } = await initOtp({
@@ -5009,7 +5008,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
               }),
             () => logout(),
             callbacks,
-            "Failed to add email"
+            "Failed to add email",
           );
         }
       } catch (error) {
@@ -5019,7 +5018,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         throw new TurnkeyError(
           "Failed to initialize OTP for email verification.",
           TurnkeyErrorCodes.INIT_OTP_ERROR,
-          error
+          error,
         );
       }
     },
@@ -5035,7 +5034,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       verifyOtp,
       updateUserEmail,
       callbacks,
-    ]
+    ],
   );
 
   const handleAddPhoneNumber = useCallback(
@@ -5050,13 +5049,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
 
       if (!masterConfig) {
         throw new TurnkeyError(
           "Config is not ready yet!",
-          TurnkeyErrorCodes.CONFIG_NOT_INITIALIZED
+          TurnkeyErrorCodes.CONFIG_NOT_INITIALIZED,
         );
       }
 
@@ -5065,14 +5064,14 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
       const userId = params?.userId || s?.userId;
       if (!userId) {
         throw new TurnkeyError(
           "A user ID is required to add a phone number.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
 
@@ -5124,7 +5123,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
               }),
             () => logout(),
             callbacks,
-            "Failed to add phone number"
+            "Failed to add phone number",
           );
         } else {
           const { otpId, otpEncryptionTargetBundle } = await initOtp({
@@ -5182,7 +5181,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
               }),
             () => logout(),
             callbacks,
-            "Failed to add phone number"
+            "Failed to add phone number",
           );
         }
       } catch (error) {
@@ -5192,7 +5191,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         throw new TurnkeyError(
           "Failed to initialize OTP for sms verification.",
           TurnkeyErrorCodes.INIT_OTP_ERROR,
-          error
+          error,
         );
       }
     },
@@ -5208,7 +5207,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       initOtp,
       verifyOtp,
       updateUserPhoneNumber,
-    ]
+    ],
   );
 
   const handleRemovePasskey = useCallback(
@@ -5225,7 +5224,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
 
       const s = await getSession();
@@ -5233,7 +5232,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
 
@@ -5265,10 +5264,10 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           }),
         () => logout(),
         callbacks,
-        "Failed to remove passkey"
+        "Failed to remove passkey",
       );
     },
-    [client, getSession, pushPage, logout, callbacks]
+    [client, getSession, pushPage, logout, callbacks],
   );
 
   const handleAddPasskey = useCallback(
@@ -5283,7 +5282,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
 
       const s = await getSession();
@@ -5291,14 +5290,14 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
       const userId = params?.userId || s?.userId;
       if (!userId) {
         throw new TurnkeyError(
           "A user ID or a valid session is required to add a passkey.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
       try {
@@ -5333,11 +5332,11 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         throw new TurnkeyError(
           "Failed to add passkey in handler.",
           TurnkeyErrorCodes.ADD_PASSKEY_ERROR,
-          error
+          error,
         );
       }
     },
-    [client, getSession, addPasskey, pushPage, closeModal]
+    [client, getSession, addPasskey, pushPage, closeModal],
   );
 
   const handleRemoveOauthProvider = useCallback(
@@ -5353,14 +5352,14 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const s = await getSession();
       const organizationId = params?.organizationId || s?.organizationId;
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
       try {
@@ -5389,8 +5388,8 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
               reject(
                 new TurnkeyError(
                   "User canceled the remove OAuth provider process.",
-                  TurnkeyErrorCodes.USER_CANCELED
-                )
+                  TurnkeyErrorCodes.USER_CANCELED,
+                ),
               ),
           });
         });
@@ -5401,11 +5400,11 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         throw new TurnkeyError(
           "Failed to remove OAuth provider in handler.",
           TurnkeyErrorCodes.REMOVE_OAUTH_PROVIDER_ERROR,
-          error
+          error,
         );
       }
     },
-    [client, getSession, pushPage]
+    [client, getSession, pushPage],
   );
 
   const handleAddOauthProvider = useCallback(
@@ -5413,7 +5412,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       const {
         providerName,
@@ -5453,14 +5452,14 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
       const userId = params?.userId || s?.userId;
       if (!userId) {
         throw new TurnkeyError(
           "A user ID is required to add an OAuth provider.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
       return new Promise<void>(async (resolve, reject) => {
@@ -5542,7 +5541,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
             default: {
               throw new TurnkeyError(
                 `Unsupported OAuth provider: ${providerName}`,
-                TurnkeyErrorCodes.NOT_FOUND
+                TurnkeyErrorCodes.NOT_FOUND,
               );
             }
           }
@@ -5562,23 +5561,23 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       handleGoogleOauth,
       handleAppleOauth,
       handleFacebookOauth,
-    ]
+    ],
   );
 
   const handleConnectExternalWallet = useCallback(
     async (
-      params?: HandleConnectExternalWalletParams
+      params?: HandleConnectExternalWalletParams,
     ): Promise<{ type: "connect" | "disconnect"; account: WalletAccount }> => {
       const { successPageDuration = 2000 } = params || {};
       if (!client)
         throw new TurnkeyError(
           "Client is not initialized.",
-          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED
+          TurnkeyErrorCodes.CLIENT_NOT_INITIALIZED,
         );
       if (!masterConfig?.walletConfig?.features?.connecting) {
         throw new TurnkeyError(
           "Wallet connecting is not enabled.",
-          TurnkeyErrorCodes.FEATURE_NOT_ENABLED
+          TurnkeyErrorCodes.FEATURE_NOT_ENABLED,
         );
       }
 
@@ -5590,7 +5589,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
               successPageDuration={successPageDuration}
               onSuccess={(
                 type: "connect" | "disconnect",
-                account: WalletAccount
+                account: WalletAccount,
               ) => {
                 resolve({ type, account });
               }}
@@ -5600,13 +5599,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
             reject(
               new TurnkeyError(
                 "User canceled the connect wallet process.",
-                TurnkeyErrorCodes.USER_CANCELED
-              )
+                TurnkeyErrorCodes.USER_CANCELED,
+              ),
             ),
         });
       });
     },
-    [client, masterConfig, fetchWalletProviders, pushPage]
+    [client, masterConfig, fetchWalletProviders, pushPage],
   );
 
   const handleRemoveUserEmail = useCallback(
@@ -5617,7 +5616,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
 
@@ -5645,8 +5644,8 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
               reject(
                 new TurnkeyError(
                   "User canceled the remove email process.",
-                  TurnkeyErrorCodes.USER_CANCELED
-                )
+                  TurnkeyErrorCodes.USER_CANCELED,
+                ),
               ),
           });
         });
@@ -5657,11 +5656,11 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         throw new TurnkeyError(
           "Failed to remove user email.",
           TurnkeyErrorCodes.UPDATE_USER_EMAIL_ERROR,
-          error
+          error,
         );
       }
     },
-    [getSession, pushPage]
+    [getSession, pushPage],
   );
 
   const handleRemoveUserPhoneNumber = useCallback(
@@ -5672,7 +5671,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
 
@@ -5700,8 +5699,8 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
               reject(
                 new TurnkeyError(
                   "User canceled the remove phone number process.",
-                  TurnkeyErrorCodes.USER_CANCELED
-                )
+                  TurnkeyErrorCodes.USER_CANCELED,
+                ),
               ),
           });
         });
@@ -5712,11 +5711,11 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         throw new TurnkeyError(
           "Failed to remove user phone number.",
           TurnkeyErrorCodes.UPDATE_USER_PHONE_NUMBER_ERROR,
-          error
+          error,
         );
       }
     },
-    [getSession, pushPage]
+    [getSession, pushPage],
   );
 
   const handleSendTransaction = useCallback(
@@ -5727,7 +5726,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
 
@@ -5823,8 +5822,8 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
             reject(
               new TurnkeyError(
                 "User canceled the transaction.",
-                TurnkeyErrorCodes.USER_CANCELED
-              )
+                TurnkeyErrorCodes.USER_CANCELED,
+              ),
             ),
           content: (
             <SendTransactionPage
@@ -5938,7 +5937,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
 
@@ -6025,7 +6024,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
                 onRampWindow = window.open(
                   result.onRampUrl.toString(),
                   "_blank",
-                  `width=${popupWidth},height=${popupHeight},top=${top},left=${left},scrollbars=yes,resizable=yes`
+                  `width=${popupWidth},height=${popupHeight},top=${top},left=${left},scrollbars=yes,resizable=yes`,
                 );
                 if (!onRampWindow)
                   throw new Error("Failed to open On Ramp window.");
@@ -6072,7 +6071,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
                     const status = resp?.transactionStatus;
                     if (
                       ["COMPLETED", "FAILED", "CANCELLED"].includes(
-                        status ?? ""
+                        status ?? "",
                       )
                     ) {
                       cleanup();
@@ -6122,13 +6121,13 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
             reject(
               new TurnkeyError(
                 "User canceled the onramp process.",
-                TurnkeyErrorCodes.USER_CANCELED
-              )
+                TurnkeyErrorCodes.USER_CANCELED,
+              ),
             ),
         });
       });
     },
-    [getSession, client, pushPage]
+    [getSession, client, pushPage],
   );
 
   const handleVerifyAppProofs = useCallback(
@@ -6139,7 +6138,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (!organizationId) {
         throw new TurnkeyError(
           "A session or passed in organization ID is required.",
-          TurnkeyErrorCodes.INVALID_REQUEST
+          TurnkeyErrorCodes.INVALID_REQUEST,
         );
       }
 
@@ -6168,8 +6167,8 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
               reject(
                 new TurnkeyError(
                   "User canceled the verify app proofs process.",
-                  TurnkeyErrorCodes.USER_CANCELED
-                )
+                  TurnkeyErrorCodes.USER_CANCELED,
+                ),
               ),
           });
         });
@@ -6180,11 +6179,11 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
         throw new TurnkeyError(
           "Failed to verify app proofs.",
           TurnkeyErrorCodes.VERIFY_APP_PROOFS_ERROR,
-          error
+          error,
         );
       }
     },
-    [getSession, pushPage]
+    [getSession, pushPage],
   );
 
   useEffect(() => {
@@ -6199,7 +6198,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
           // Only fetch the proxy auth config if we have an authProxyId and the autoFetchWalletKitConfig param is enabled or not passed in.
           proxyAuthConfig = await getAuthProxyConfig(
             config.authProxyConfigId!, // Can assert safely. See shouldFetchWalletKitConfig definition.
-            config.authProxyUrl
+            config.authProxyUrl,
           );
           proxyAuthConfigRef.current = proxyAuthConfig;
         }
@@ -6241,11 +6240,11 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
    */
   const debouncedRefreshWallets = useDebouncedCallback(
     maybeRefreshWallets,
-    100
+    100,
   );
   const debouncedFetchWalletProviders = useDebouncedCallback(
     fetchWalletProviders,
-    100
+    100,
   );
 
   useEffect(() => {
