+++ packages/react-wallet-kit/src/providers/client/Types.tsx
@@ -341,7 +341,7 @@ export interface ClientContextType
    * @return A void promise.
    */
   handleExportPrivateKey: (
-    params: HandleExportPrivateKeyParams
+    params: HandleExportPrivateKeyParams,
   ) => Promise<void>;
 
   /**
@@ -368,7 +368,7 @@ export interface ClientContextType
    */
 
   handleExportWalletAccount: (
-    params: HandleExportWalletAccountParams
+    params: HandleExportWalletAccountParams,
   ) => Promise<void>;
 
   /**
@@ -416,7 +416,7 @@ export interface ClientContextType
    * @returns A promise that resolves to the new private key's ID.
    */
   handleImportPrivateKey: (
-    params?: HandleImportPrivateKeyParams
+    params?: HandleImportPrivateKeyParams,
   ) => Promise<string>;
 
   /**
@@ -442,7 +442,7 @@ export interface ClientContextType
    * @throws {TurnkeyError} If the client is not initialized, no active session is found, or if there is an error updating the user name.
    */
   handleUpdateUserEmail: (
-    params?: HandleUpdateUserEmailParams
+    params?: HandleUpdateUserEmailParams,
   ) => Promise<string>;
 
   /**
@@ -469,7 +469,7 @@ export interface ClientContextType
    * @throws {TurnkeyError} If the client is not initialized, no active session is found, SMS OTP is not enabled, or if there is an error updating the phone number.
    */
   handleUpdateUserPhoneNumber: (
-    params?: HandleUpdateUserPhoneNumberParams
+    params?: HandleUpdateUserPhoneNumberParams,
   ) => Promise<string>;
   /**
    * Handles the update user name flow.
@@ -493,7 +493,7 @@ export interface ClientContextType
    * @throws {TurnkeyError} If the client is not initialized, no active session is found, or if there is an error updating the email.
    */
   handleUpdateUserName: (
-    params?: HandleUpdateUserNameParams
+    params?: HandleUpdateUserNameParams,
   ) => Promise<string>;
   /**
    * Handles the add user email flow.
@@ -543,7 +543,7 @@ export interface ClientContextType
    * @throws {TurnkeyError} If the client is not initialized, no active session is found, or if there is an error adding the phone number.
    */
   handleAddPhoneNumber: (
-    params?: HandleAddPhoneNumberParams
+    params?: HandleAddPhoneNumberParams,
   ) => Promise<string>;
 
   /**
@@ -572,7 +572,7 @@ export interface ClientContextType
    * @throws {TurnkeyError} If the client is not initialized, no active session is found, or if there is an error adding the provider.
    */
   handleAddOauthProvider: (
-    params: HandleAddOauthProviderParams
+    params: HandleAddOauthProviderParams,
   ) => Promise<void>;
 
   /**
@@ -598,7 +598,7 @@ export interface ClientContextType
    * if there is an error removing the provider, or if the user cancels the action.
    */
   handleRemoveOauthProvider: (
-    params: HandleRemoveOauthProviderParams
+    params: HandleRemoveOauthProviderParams,
   ) => Promise<string[]>;
 
   /**
@@ -669,7 +669,7 @@ export interface ClientContextType
    * @throws {TurnkeyError} If the client is not initialized, if there is an error during the signing process, or if the user cancels the action.
    */
   handleSignMessage: (
-    params: HandleSignMessageParams
+    params: HandleSignMessageParams,
   ) => Promise<v1SignRawPayloadResult>;
 
   /**
@@ -690,7 +690,7 @@ export interface ClientContextType
    * @throws {TurnkeyError} If the client is not initialized or if the user cancels the action.
    */
   handleConnectExternalWallet: (
-    params?: HandleConnectExternalWalletParams
+    params?: HandleConnectExternalWalletParams,
   ) => Promise<{
     type: "connect" | "disconnect";
     account: WalletAccount;
@@ -713,7 +713,7 @@ export interface ClientContextType
    * if the user cancels the action, or if there is an error during the removal process.
    */
   handleRemoveUserEmail: (
-    params?: HandleRemoveUserEmailParams
+    params?: HandleRemoveUserEmailParams,
   ) => Promise<string>;
 
   /**
@@ -733,7 +733,7 @@ export interface ClientContextType
    * if the user cancels the action, or if there is an error during the removal process.
    */
   handleRemoveUserPhoneNumber: (
-    params?: HandleRemoveUserPhoneNumberParams
+    params?: HandleRemoveUserPhoneNumberParams,
   ) => Promise<string>;
 
   /**
@@ -823,10 +823,20 @@ export interface ClientContextType
     params: HandleSendErc20TransferParams,
   ) => Promise<void>;
   
+  /**
+   * Fetches WalletConnect apps from the WalletConnect 'v3/wallets' api and builds Turnkey WalletProvider objects.
+   *
+   * - This function retrieves WalletConnect providers and updates the 'walletConnectApps' state with the fetched providers.
+   * - It throws an error if the client is not initialized or if the WalletConnect project ID is not configured.
+   * - These providers will be built with an empty 'WalletRpcProvider' object. You will need to use the WalletConnect provider from the 'walletProviders' state to use WalletConnect functionality.
+   *
+   * @returns A promise that resolves to an array of WalletProvider objects representing the WalletConnect apps.
+   * @throws {TurnkeyError} If the client is not initialized or if the WalletConnect project ID is not configured.
+   */
   fetchAndBuildWalletConnectApps: () => Promise<WalletProvider[]>;
 }
 
 /** @internal */
 export const ClientContext = createContext<ClientContextType | undefined>(
-  undefined
+  undefined,
 );
