+++ packages/react-wallet-kit/src/utils/utils.ts
@@ -47,7 +47,7 @@ export const authErrors = {
 
 export const useDebouncedCallback = <T extends (...args: any[]) => void>(
   fn: T,
-  wait = 100
+  wait = 100,
 ): T => {
   const timer = useRef<ReturnType<typeof setTimeout> | null>(null);
   const fnRef = useRef(fn);
@@ -62,7 +62,7 @@ export const useDebouncedCallback = <T extends (...args: any[]) => void>(
         timer.current = null;
       }, wait);
     },
-    [wait]
+    [wait],
   ) as T;
 };
 
@@ -75,7 +75,7 @@ export async function withTurnkeyErrorHandling<T>(
   sessionExpireFn?: () => Promise<void>,
   callbacks?: { onError?: (error: TurnkeyError) => void },
   fallbackMessage = "An unknown error occurred",
-  fallbackCode = TurnkeyErrorCodes.UNKNOWN
+  fallbackCode = TurnkeyErrorCodes.UNKNOWN,
 ): Promise<T> {
   try {
     return await fn();
@@ -95,7 +95,7 @@ export async function withTurnkeyErrorHandling<T>(
           throw new TurnkeyError(
             "SESSION_EXPIRED received without sessionExpireFn handler",
             TurnkeyErrorCodes.INTERNAL_ERROR,
-            error
+            error,
           );
         }
       }
@@ -118,10 +118,10 @@ export async function withTurnkeyErrorHandling<T>(
 // Custom hook to get the current screen size
 export function useScreenSize() {
   const [width, setWidth] = useState(
-    typeof window !== "undefined" ? window.innerWidth : 1024
+    typeof window !== "undefined" ? window.innerWidth : 1024,
   );
   const [height, setHeight] = useState(
-    typeof window !== "undefined" ? window.innerHeight : 768
+    typeof window !== "undefined" ? window.innerHeight : 768,
   );
 
   useEffect(() => {
@@ -150,12 +150,12 @@ export function isWalletConnect(wallet: WalletProvider): boolean {
 // Finds and returns the WalletConnect provider for the specified chain
 export function findWalletConnectProvider(
   walletProviders: WalletProvider[],
-  chain: Chain
+  chain: Chain,
 ): WalletProvider | undefined {
   return walletProviders.find(
     (p) =>
       p.interfaceType === WalletInterfaceType.WalletConnect &&
-      p.chainInfo.namespace === chain
+      p.chainInfo.namespace === chain,
   );
 }
 
@@ -193,7 +193,7 @@ export function useWalletProviderState(initialState: WalletProvider[] = []) {
       }
       // we do nothing if the wallet providers are the same
     },
-    []
+    [],
   );
 
   return [walletProviders, updateWalletProviders] as const;
@@ -201,11 +201,11 @@ export function useWalletProviderState(initialState: WalletProvider[] = []) {
 
 export function mergeWalletsWithoutDuplicates(
   existingWallets: Wallet[],
-  newWallets: Wallet[]
+  newWallets: Wallet[],
 ): Wallet[] {
   const existingWalletIds = new Set(existingWallets.map((w) => w.walletId));
   const uniqueNewWallets = newWallets.filter(
-    (w) => !existingWalletIds.has(w.walletId)
+    (w) => !existingWalletIds.has(w.walletId),
   );
   return [...existingWallets, ...uniqueNewWallets];
 }
@@ -225,7 +225,7 @@ export async function buildWalletConnectProviders(params: {
 
     // Check if the app supports Ethereum
     const supportsEthereum = app.chains.some((chain) =>
-      chain.startsWith("eip155:")
+      chain.startsWith("eip155:"),
     );
     if (supportsEthereum) {
       supportedChains.push(Chain.Ethereum);
@@ -233,7 +233,7 @@ export async function buildWalletConnectProviders(params: {
 
     // Check if the app supports Solana
     const supportsSolana = app.chains.some((chain) =>
-      chain.startsWith("solana:")
+      chain.startsWith("solana:"),
     );
     if (supportsSolana) {
       supportedChains.push(Chain.Solana);
